+++ WaterWatch WatchKit Extension/Views/GoalOuncesView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 struct GoalOuncesView: View {
     @Binding var isPresented: Bool
-    @Binding var goalMilliliters: Int
+    @Binding var goalMilliliters: Double
     @State private var sampleSize = 100.0
     
     var body: some View {
@@ -23,14 +23,14 @@ struct GoalOuncesView: View {
             Button(action: {
                 let measurement = Measurement(value: self.sampleSize, unit: UnitVolume.fluidOunces)
                 let milliliters = measurement.converted(to: .milliliters).value
-                self.goalMilliliters = Int(milliliters)
+                self.goalMilliliters = milliliters
                 self.isPresented = false
             }) {
                 Text("Set")
                     .font(.system(size: 20, weight: Font.Weight.regular, design: Font.Design.rounded))
             }
         }.onAppear() {
-            let measurement = Measurement(value: Double(self.goalMilliliters), unit: UnitVolume.milliliters)
+            let measurement = Measurement(value: self.goalMilliliters, unit: UnitVolume.milliliters)
             let goalOunces = measurement.converted(to: .fluidOunces).value
             self.sampleSize = goalOunces
         }
